OEL-448: Refactor corporate block into two.
